Extract CommonPredicatesExpressionRewriter
