+++ presto-main/src/main/java/com/facebook/presto/sql/planner/optimizations/SimplifyExpressions.java
@@ -17,7 +17,6 @@ import com.facebook.presto.Session;
 import com.facebook.presto.metadata.Metadata;
 import com.facebook.presto.spi.type.Type;
 import com.facebook.presto.sql.parser.SqlParser;
-import com.facebook.presto.sql.planner.DeterminismEvaluator;
 import com.facebook.presto.sql.planner.ExpressionInterpreter;
 import com.facebook.presto.sql.planner.LiteralInterpreter;
 import com.facebook.presto.sql.planner.NoOpSymbolResolver;
@@ -32,34 +31,20 @@ import com.facebook.presto.sql.planner.plan.SimplePlanRewriter;
 import com.facebook.presto.sql.planner.plan.TableScanNode;
 import com.facebook.presto.sql.planner.plan.ValuesNode;
 import com.facebook.presto.sql.tree.Expression;
-import com.facebook.presto.sql.tree.ExpressionRewriter;
-import com.facebook.presto.sql.tree.ExpressionTreeRewriter;
-import com.facebook.presto.sql.tree.LogicalBinaryExpression;
 import com.facebook.presto.sql.tree.NodeRef;
 import com.facebook.presto.sql.tree.NullLiteral;
 import com.facebook.presto.sql.tree.SymbolReference;
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableSet;
-import com.google.common.collect.Sets;
 
-import java.util.Collection;
-import java.util.List;
 import java.util.Map;
-import java.util.Set;
 
-import static com.facebook.presto.sql.ExpressionUtils.combinePredicates;
-import static com.facebook.presto.sql.ExpressionUtils.extractPredicates;
 import static com.facebook.presto.sql.analyzer.ExpressionAnalyzer.getExpressionTypes;
+import static com.facebook.presto.sql.planner.iterative.rule.ExtractCommonPredicatesExpressionRewriter.extractCommonPredicates;
 import static com.facebook.presto.sql.planner.iterative.rule.PushDownNegationsExpressionRewriter.pushDownNegations;
 import static com.facebook.presto.sql.tree.BooleanLiteral.FALSE_LITERAL;
 import static com.facebook.presto.sql.tree.BooleanLiteral.TRUE_LITERAL;
-import static com.facebook.presto.sql.tree.LogicalBinaryExpression.Type.OR;
-import static com.google.common.collect.ImmutableList.toImmutableList;
 import static java.util.Collections.emptyList;
-import static java.util.Collections.emptySet;
 import static java.util.Objects.requireNonNull;
-import static java.util.stream.Collectors.toList;
-import static java.util.stream.Collectors.toSet;
 
 public class SimplifyExpressions
         implements PlanOptimizer
@@ -167,158 +152,10 @@ public class SimplifyExpressions
                 return expression;
             }
             expression = pushDownNegations(expression);
-            expression = ExpressionTreeRewriter.rewriteWith(new ExtractCommonPredicatesExpressionRewriter(), expression, NodeContext.ROOT_NODE);
+            expression = extractCommonPredicates(expression);
             Map<NodeRef<Expression>, Type> expressionTypes = getExpressionTypes(session, metadata, sqlParser, types, expression, emptyList() /* parameters already replaced */);
             ExpressionInterpreter interpreter = ExpressionInterpreter.expressionOptimizer(expression, metadata, session, expressionTypes);
             return LiteralInterpreter.toExpression(interpreter.optimize(NoOpSymbolResolver.INSTANCE), expressionTypes.get(NodeRef.of(expression)));
-        }
-    }
-
-    private enum NodeContext
-    {
-        ROOT_NODE,
-        NOT_ROOT_NODE;
-
-        boolean isRootNode()
-        {
-            return this == ROOT_NODE;
-        }
-    }
-
-    private static class ExtractCommonPredicatesExpressionRewriter
-            extends ExpressionRewriter<NodeContext>
-    {
-        @Override
-        public Expression rewriteExpression(Expression node, NodeContext context, ExpressionTreeRewriter<NodeContext> treeRewriter)
-        {
-            if (context.isRootNode()) {
-                return treeRewriter.rewrite(node, NodeContext.NOT_ROOT_NODE);
-            }
-
-            return null;
-        }
-
-        @Override
-        public Expression rewriteLogicalBinaryExpression(LogicalBinaryExpression node, NodeContext context, ExpressionTreeRewriter<NodeContext> treeRewriter)
-        {
-            Expression expression = combinePredicates(
-                    node.getType(),
-                    extractPredicates(node.getType(), node).stream()
-                            .map(subExpression -> treeRewriter.rewrite(subExpression, NodeContext.NOT_ROOT_NODE))
-                            .collect(toImmutableList()));
-
-            if (!(expression instanceof LogicalBinaryExpression)) {
-                return expression;
-            }
-
-            Expression simplified = extractCommonPredicates((LogicalBinaryExpression) expression);
-
-            // Prefer AND LogicalBinaryExpression at the root if possible
-            if (context.isRootNode() && simplified instanceof LogicalBinaryExpression && ((LogicalBinaryExpression) simplified).getType() == OR) {
-                return distributeIfPossible((LogicalBinaryExpression) simplified);
-            }
-
-            return simplified;
-        }
-
-        private static Expression extractCommonPredicates(LogicalBinaryExpression node)
-        {
-            List<List<Expression>> subPredicates = getSubPredicates(node);
-
-            Set<Expression> commonPredicates = ImmutableSet.copyOf(subPredicates.stream()
-                    .map(ExtractCommonPredicatesExpressionRewriter::filterDeterministicPredicates)
-                    .reduce(Sets::intersection)
-                    .orElse(emptySet()));
-
-            List<List<Expression>> uncorrelatedSubPredicates = subPredicates.stream()
-                    .map(predicateList -> removeAll(predicateList, commonPredicates))
-                    .collect(toImmutableList());
-
-            LogicalBinaryExpression.Type flippedNodeType = node.getType().flip();
-
-            List<Expression> uncorrelatedPredicates = uncorrelatedSubPredicates.stream()
-                    .map(predicate -> combinePredicates(flippedNodeType, predicate))
-                    .collect(toImmutableList());
-            Expression combinedUncorrelatedPredicates = combinePredicates(node.getType(), uncorrelatedPredicates);
-
-            return combinePredicates(flippedNodeType, ImmutableList.<Expression>builder()
-                    .addAll(commonPredicates)
-                    .add(combinedUncorrelatedPredicates)
-                    .build());
-        }
-
-        private static List<List<Expression>> getSubPredicates(LogicalBinaryExpression expression)
-        {
-            return extractPredicates(expression.getType(), expression).stream()
-                    .map(predicate -> predicate instanceof LogicalBinaryExpression ?
-                            extractPredicates((LogicalBinaryExpression) predicate) : ImmutableList.of(predicate))
-                    .collect(toImmutableList());
-        }
-
-        /**
-         * Applies the boolean distributive property.
-         * <p>
-         * For example:
-         * (A & B) | (C & D) => (A | C) & (A | D) & (B | C) & (B | D)
-         * <p>
-         * Returns the original expression if the expression is non-deterministic or if the distribution will
-         * expand the expression by too much.
-         */
-        private static Expression distributeIfPossible(LogicalBinaryExpression expression)
-        {
-            if (!DeterminismEvaluator.isDeterministic(expression)) {
-                // Do not distribute boolean expressions if there are any non-deterministic elements
-                // TODO: This can be optimized further if non-deterministic elements are not repeated
-                return expression;
-            }
-            List<Set<Expression>> subPredicates = getSubPredicates(expression).stream()
-                    .map(ImmutableSet::copyOf)
-                    .collect(toList());
-
-            int originalBaseExpressions = subPredicates.stream()
-                    .mapToInt(Set::size)
-                    .sum();
-
-            int newBaseExpressions;
-            try {
-                newBaseExpressions = Math.multiplyExact(subPredicates.stream()
-                        .mapToInt(Set::size)
-                        .reduce(Math::multiplyExact)
-                        .getAsInt(), subPredicates.size());
-            }
-            catch (ArithmeticException e) {
-                // Integer overflow from multiplication means there are too many expressions
-                return expression;
-            }
-
-            if (newBaseExpressions > originalBaseExpressions * 2) {
-                // Do not distribute boolean expressions if it would create 2x more base expressions
-                // (e.g. A, B, C, D from the above example). This is just an arbitrary heuristic to
-                // avoid cross product expression explosion.
-                return expression;
-            }
-
-            Set<List<Expression>> crossProduct = Sets.cartesianProduct(subPredicates);
-
-            return combinePredicates(
-                    expression.getType().flip(),
-                    crossProduct.stream()
-                            .map(expressions -> combinePredicates(expression.getType(), expressions))
-                            .collect(toImmutableList()));
-        }
-
-        private static Set<Expression> filterDeterministicPredicates(List<Expression> predicates)
-        {
-            return predicates.stream()
-                    .filter(DeterminismEvaluator::isDeterministic)
-                    .collect(toSet());
-        }
-
-        private static <T> List<T> removeAll(Collection<T> collection, Collection<T> elementsToRemove)
-        {
-            return collection.stream()
-                    .filter(element -> !elementsToRemove.contains(element))
-                    .collect(toImmutableList());
         }
     }
 }
